requirements.txt
numpy
scipy
matplotlib
seaborn

cerberus
pytest

trimesh
pyglet
Rtree

toml
pandas

tqdm

cupy-rocm-4-3
